+++ packages/create-instantsearch-app/src/templates/Vue InstantSearch/src/App.vue
@@ -18,6 +18,7 @@
       <ais-instant-search
         :search-client="searchClient"
         index-name="{{indexName}}"
+        {{#if flags.insights}}insights{{/if}}
       >
         <ais-configure :hits-per-page.camel="8" />
         <div class="search-panel">