+++ packages/instantsearch.js/src/__tests__/common.test.tsx
@@ -9,6 +9,8 @@ import {
   createMenuTests,
   createInfiniteHitsTests,
   createHitsTests,
+  createRangeInputTests,
+  createInstantSearchTests,
 } from '@instantsearch/tests';
 
 import instantsearch from '../index.es';
@@ -22,6 +24,7 @@ import {
   searchBox,
   hits,
   index,
+  rangeInput,
 } from '../widgets';
 
 createHierarchicalMenuTests(({ instantSearchOptions, widgetParams }) => {
@@ -242,4 +245,40 @@ createHitsTests(({ instantSearchOptions, widgetParams }) => {
        */
     })
     .start();
+});
+
+createRangeInputTests(({ instantSearchOptions, widgetParams }) => {
+  instantsearch(instantSearchOptions)
+    .addWidgets([
+      rangeInput({
+        container: document.body.appendChild(document.createElement('div')),
+        ...widgetParams,
+      }),
+    ])
+    .on('error', () => {
+      /*
+       * prevent rethrowing InstantSearch errors, so tests can be asserted.
+       * IRL this isn't needed, as the error doesn't stop execution.
+       */
+    })
+    .start();
+});
+
+createInstantSearchTests(({ instantSearchOptions }) => {
+  instantsearch(instantSearchOptions)
+    .on('error', () => {
+      /*
+       * prevent rethrowing InstantSearch errors, so tests can be asserted.
+       * IRL this isn't needed, as the error doesn't stop execution.
+       */
+    })
+    .start();
+
+  return {
+    algoliaAgents: [
+      `instantsearch.js (${
+        require('../../../instantsearch.js/package.json').version
+      })`,
+    ],
+  };
 });