+++ packages/react-instantsearch-hooks-web/src/ui/InfiniteHits.tsx
@@ -116,6 +116,9 @@ export function InfiniteHits<THit extends Hit>({
             onClick={() => {
               sendEvent('click:internal', hit, 'Hit Clicked');
             }}
+            onAuxClick={() => {
+              sendEvent('click:internal', hit, 'Hit Clicked');
+            }}
           >
             <HitComponent hit={hit} sendEvent={sendEvent} />
           </li>