+++ packages/react-instantsearch-hooks/src/lib/useInstantSearchApi.ts
@@ -21,6 +21,8 @@ const defaultUserAgents = [
   `react-instantsearch-hooks (${version})`,
 ];
 const serverUserAgent = `react-instantsearch-server (${version})`;
+const nextUserAgent = (nextVersion?: string) =>
+  nextVersion ? `next.js (${nextVersion})` : null;
 
 export type UseInstantSearchApiProps<
   TUiState extends UiState,
@@ -102,6 +104,7 @@ export function useInstantSearchApi<TUiState extends UiState, TRouteState>(
     addAlgoliaAgents(props.searchClient, [
       ...defaultUserAgents,
       serverContext && serverUserAgent,
+      nextUserAgent(getNextVersion()),
     ]);
 
     // On the server, we start the search early to compute the search parameters.
@@ -261,4 +264,17 @@ Please check its usage instructions: https://github.com/algolia/instantsearch/tr
 You can ignore this warning if you are using a custom router that suits your needs, it won't be outputted in production builds.`
     );
   }
+}
+
+/**
+ * Gets the version of Next.js if it is available in the `window` object,
+ * otherwise it returns the NEXT_RUNTIME environment variable (in SSR),
+ * which is either `nodejs` or `edge`.
+ */
+function getNextVersion() {
+  return (
+    (typeof window !== 'undefined' &&
+      ((window as any).next?.version as string | undefined)) ||
+    (typeof process !== 'undefined' ? process.env?.NEXT_RUNTIME : undefined)
+  );
 }