+++ tests/common/index.ts
@@ -5,3 +5,5 @@ export * from './widgets/menu';
 export * from './widgets/pagination';
 export * from './widgets/infinite-hits';
 export * from './widgets/hits';
+export * from './widgets/range-input';
+export * from './widgets/instantsearch';