+++ tests/utils/index.ts
@@ -2,6 +2,7 @@ export * from './castToJestMock';
 export * from './createInstantSearchSpy';
 export * from './enzyme';
 export * from './InstantSearchHooksTestWrapper';
+export * from './normalizeSnapshot';
 export * from './runAllMicroTasks';
 export * from './wait';
 export * from './widgetSnapshotSerializer';