+++ packages/create-instantsearch-app/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.4.0](https://github.com/algolia/instantsearch/compare/create-instantsearch-app@6.3.1...create-instantsearch-app@6.4.0) (2023-05-30)
+
+
+### Bug Fixes
+
+* **dependencies:** remove fixed instantsearch.js dependency from react instantsearch hooks templates ([#5636](https://github.com/algolia/instantsearch/issues/5636)) ([9f41d7e](https://github.com/algolia/instantsearch/commit/9f41d7e702bfa482b229846e67ee4c7619f756cf))
+* **dependencies:** update metalsmith packages ([#5642](https://github.com/algolia/instantsearch/issues/5642)) ([e001aa9](https://github.com/algolia/instantsearch/commit/e001aa9a06bab28777906901f4f4ea007c92a2a1))
+* **js:** use html templates ([#5644](https://github.com/algolia/instantsearch/issues/5644)) ([4bcc426](https://github.com/algolia/instantsearch/commit/4bcc4267cdbecdb9274b0fe87e0a4f8701ec6475))
+
+
+### Features
+
+* **templates:** add "insights" flag if supported ([#5639](https://github.com/algolia/instantsearch/issues/5639)) ([212ad4f](https://github.com/algolia/instantsearch/commit/212ad4fd2d40c525dfc5e49d0c77fc3013275c38))
+
+
+
+
+
 ## [6.3.1](https://github.com/algolia/instantsearch/compare/create-instantsearch-app@6.3.0...create-instantsearch-app@6.3.1) (2023-04-24)
 
 **Note:** Version bump only for package create-instantsearch-app