+++ packages/react-instantsearch-native/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.40.1](https://github.com/algolia/instantsearch.js/compare/react-instantsearch-native@6.40.0...react-instantsearch-native@6.40.1) (2023-06-20)
+
+**Note:** Version bump only for package react-instantsearch-native
+
+
+
+
+
+## [6.39.2](https://github.com/algolia/instantsearch.js/compare/react-instantsearch-native@6.39.1...react-instantsearch-native@6.39.2) (2023-05-16)
+
+**Note:** Version bump only for package react-instantsearch-native
+
+
+
+
+
 ## [6.39.1](https://github.com/algolia/instantsearch.js/compare/react-instantsearch-native@6.38.3...react-instantsearch-native@6.39.1) (2023-01-26)
 
 **Note:** Version bump only for package react-instantsearch-native