+++ packages/vue-instantsearch/CHANGELOG.md
@@ -3,6 +3,56 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.10.3](https://github.com/algolia/instantsearch.js/compare/vue-instantsearch@4.10.2...vue-instantsearch@4.10.3) (2023-06-20)
+
+
+### Bug Fixes
+
+* **dependencies:** update helper requirement ([#5676](https://github.com/algolia/instantsearch.js/issues/5676)) ([c289120](https://github.com/algolia/instantsearch.js/commit/c2891205c1125b1203b3b3db946d57e0fc4e4687)), closes [#5658](https://github.com/algolia/instantsearch.js/issues/5658)
+
+
+
+
+
+## [4.10.2](https://github.com/algolia/instantsearch.js/compare/vue-instantsearch@4.10.1...vue-instantsearch@4.10.2) (2023-06-13)
+
+
+### Bug Fixes
+
+* **createServerRootMixin:** forward i18n on option ([#5673](https://github.com/algolia/instantsearch/issues/5673)) ([d52b47f](https://github.com/algolia/instantsearch/commit/d52b47f58a9e723616ec59c9805c590e8079ba38))
+* **breadcrumb:** align vue component with specs ([#5672](https://github.com/algolia/instantsearch.js/issues/5672)) ([dbfed3a](https://github.com/algolia/instantsearch.js/commit/dbfed3ac25910013835c9bb36255b6dbd6d264ec))
+
+
+
+
+
+## [4.10.1](https://github.com/algolia/instantsearch.js/compare/vue-instantsearch@4.10.0...vue-instantsearch@4.10.1) (2023-05-30)
+
+
+### Bug Fixes
+
+* **insights:** send default click event when using auxiliary pointer button ([#5634](https://github.com/algolia/instantsearch.js/issues/5634)) ([7e4a216](https://github.com/algolia/instantsearch.js/commit/7e4a2162f87596a384b35c97efa51db9bb6f8973))
+
+
+
+
+
+# [4.10.0](https://github.com/algolia/instantsearch.js/compare/vue-instantsearch@4.9.0...vue-instantsearch@4.10.0) (2023-05-16)
+
+
+### Bug Fixes
+
+* **rangeinput:** allow input of numbers with precision ([#5541](https://github.com/algolia/instantsearch.js/issues/5541)) ([fb48951](https://github.com/algolia/instantsearch.js/commit/fb489513a8550528f3e2867be30fb380229ad188))
+
+
+### Features
+
+* **instantsearch:** make root indexName optional ([#5590](https://github.com/algolia/instantsearch.js/issues/5590)) ([80f309e](https://github.com/algolia/instantsearch.js/commit/80f309ed69b61534ca118b60c9c88691e0148fca))
+
+
+
+
+
 # [4.9.0](https://github.com/algolia/instantsearch.js/compare/vue-instantsearch@4.8.10...vue-instantsearch@4.9.0) (2023-04-24)